Support inline CSS in markdown posts
Make it possible to declare post-specific CSS with the following
syntax in the post markdown:

``` {.hakyll-inline-css}
div.foo {
 ...
}
```

This will be automatically hoisted to HTML page <head> section for the
particular post that declared the CSS fragment.

The main use (I'm sure there will be plenty) would be flexbox
declarations for images.